requirements.txt
numpy>=1.26
pandas>=2.2
paho-mqtt>=2.1
pytest>=8.2
